+++ src/main/java/my/custom/learning/weblearn/entity/User.java
@@ -41,7 +41,7 @@ public class User {
 
 	@NotNull(message = "The address id cannot be null.")
 	@Column(name = "address_id")
-	private Integer user_address_id;
+	private Long user_address_id;
 
 	@OneToOne
 	@JoinColumn(name = "address_id", referencedColumnName = "address_id", insertable = false, updatable = false)
@@ -59,7 +59,7 @@ public class User {
 			@Pattern(regexp = "^[A-Z]{1}[a-z]+$", message = "The first name should only contain alphabets. The first alphabet of the first name must be a capital letter.") String first_name,
 			@Pattern(regexp = "^[A-Z]{1}[a-z]+$", message = "The last name should only contain alphabets. The first alphabet of the last name must be a capital letter.") String last_name,
 			@Email(message = "The email format is not correct. ") String email, String secret_key,
-			@NotNull(message = "The phone number cannot be null.") String phone, Integer address_id, boolean is_active,
+			@NotNull(message = "The phone number cannot be null.") String phone, Long address_id, boolean is_active,
 			LocalDateTime created_at) {
 		super();
 		this.user_id = user_id;
@@ -144,11 +144,11 @@ public class User {
 		this.address = address;
 	}
 
-	public Integer getUser_address_id() {
+	public Long getUser_address_id() {
 		return user_address_id;
 	}
 
-	public void setUser_address_id(Integer user_address_id) {
+	public void setUser_address_id(Long user_address_id) {
 		this.user_address_id = user_address_id;
 	}
