+++ src/main/java/my/custom/learning/weblearn/repository/ProcuctsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import my.custom.learning.weblearn.entity.Products;
 
-public interface ProcuctsRepository extends JpaRepository<Products, Integer> {
+public interface ProcuctsRepository extends JpaRepository<Products, Long> {
 
 	// Custom query: user password is stored in the `secret_key` property on the User entity.
 	@Query("SELECT p FROM products p WHERE p.sku = :sku")
