+++ src/main/java/my/custom/learning/weblearn/repository/ProductImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import my.custom.learning.weblearn.entity.ProductImage;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
 
 }
